+++ packages/core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - ce74986: feat: Add `resolve` config support and attempt to fix plugin support
+  - @viteval/internal@0.0.3
 
 ## 0.0.2